+++ crates/rattler_solve/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [9.0.7](https://github.com/conda/rattler/compare/rattler_solve-v9.0.6...rattler_solve-v9.0.7) - 2026-09-08
+
+### Other
+
+- updated the following local packages: rattler_libsolv_c
+
 ## [9.0.6](https://github.com/conda/rattler/compare/rattler_solve-v9.0.5...rattler_solve-v9.0.6) - 2026-09-04
 
 ### Other