+++ crates/rattler_upload/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.8](https://github.com/conda/rattler/compare/rattler_upload-v0.10.7...rattler_upload-v0.10.8) - 2026-09-08
+
+### Other
+
+- updated the following local packages: rattler_solve
+
 ## [0.10.7](https://github.com/conda/rattler/compare/rattler_upload-v0.10.6...rattler_upload-v0.10.7) - 2026-09-04
 
 ### Other